Move metadata to root layout
